set ntz to 0 if sign is set for int64
The commit fixes a bug when compressing a set of int64 numbers where the
ntz is greater than 0 but the set has a negative difference. In this
circumstance the shift to trim the ntz bits was removing the sign needed
to recover negative differences during decompression. This change fixes
to issue by setting ntz to 0 if the set has a negative difference.

fixes ronanh/intcomp#18